[Feat] : titleLabel 구현 완료
